+++ processor/src/main/java/org/realityforge/arez/processor/AutorunDescriptor.java
@@ -75,7 +75,7 @@ final class AutorunDescriptor
                          GeneratorUtil.FIELD_PREFIX + getName(),
                          Modifier.FINAL,
                          Modifier.PRIVATE ).
-        addAnnotation( Nonnull.class );
+        addAnnotation( GeneratorUtil.NONNULL_CLASSNAME );
     builder.addField( field.build() );
   }