+++ processor/src/main/java/org/realityforge/arez/processor/ComputedDescriptor.java
@@ -232,7 +232,7 @@ final class ComputedDescriptor
                          GeneratorUtil.FIELD_PREFIX + getName(),
                          Modifier.FINAL,
                          Modifier.PRIVATE ).
-        addAnnotation( Nonnull.class );
+        addAnnotation( GeneratorUtil.NONNULL_CLASSNAME );
     builder.addField( field.build() );
   }