+++ processor/src/main/java/org/realityforge/arez/processor/ObservableDescriptor.java
@@ -159,7 +159,7 @@ final class ObservableDescriptor
                          GeneratorUtil.FIELD_PREFIX + getName(),
                          Modifier.FINAL,
                          Modifier.PRIVATE ).
-        addAnnotation( Nonnull.class );
+        addAnnotation( GeneratorUtil.NONNULL_CLASSNAME );
     builder.addField( field.build() );
   }